Issue #403: Minor refactor on Exception types
