+++ RepoDb/RepoDb/DataProviderFunctions/DataProviderFunctionBuilders/BaseDataProviderFunctionBuilder.cs
@@ -1,4 +1,5 @@
-﻿﻿using System;
+﻿using RepoDb.DataProviderFunctions.Exceptions;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
@@ -93,58 +94,5 @@ namespace RepoDb.DataProviderFunctions.DataProviderFunctionBuilders {
             }
         }
         #endregion
-    }
-
-    #region Exceptions
-    /// <summary>
-    /// 
-    /// </summary>
-    public class NotSupportedFunctionException : NotSupportedException {
-        /// <summary>
-        /// 
-        /// </summary>
-        public readonly string DataProviderFunctionName;
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="dataProviderFunctionName"></param>
-        public NotSupportedFunctionException(string dataProviderFunctionName) : base(string.Format("Function {0} not supported by DataProvider",
-                                                                                                    dataProviderFunctionName)) {
-            DataProviderFunctionName = dataProviderFunctionName;
-        }
-    }
-
-    /// <summary>
-    /// 
-    /// </summary>
-    public class DataProviderFunctionDecoratorException : ApplicationException {
-        /// <summary>
-        /// 
-        /// </summary>
-        public readonly string FieldName;
-        /// <summary>
-        /// 
-        /// </summary>
-        public readonly string DataProviderFunctionName;
-        /// <summary>
-        /// 
-        /// </summary>
-        public readonly Exception Exception;
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="fieldName"></param>
-        /// <param name="dataProviderFunctionName"></param>
-        /// <param name="exc"></param>
-        public DataProviderFunctionDecoratorException(string fieldName, string dataProviderFunctionName, Exception exc) :
-                base(string.Format("Decorator for DataProviderFunction {0} threw an exception while being applied to field {1}: {2}",
-                                                    dataProviderFunctionName, fieldName, exc.Message)) {
-            FieldName = fieldName;
-            DataProviderFunctionName = dataProviderFunctionName;
-            Exception = exc;
-        }
-    }
+    }    
 }
-    #endregion Exceptions
