+++ RepoDb/RepoDb/RepoDb.csproj
@@ -74,6 +74,8 @@
     <Compile Include="DataProviderFunctions\DataProviderFunctionBuilders\IDataProviderFunctionBuilder.cs" />
     <Compile Include="DataProviderFunctions\DataProviderFunctionBuilders\SqlServerFunctionBuilder.cs" />
     <Compile Include="DataProviderFunctions\DataProviderFunctionExtractor.cs" />
+    <Compile Include="DataProviderFunctions\Exceptions\DataProviderFunctionDecoratorException.cs" />
+    <Compile Include="DataProviderFunctions\Exceptions\NotSupportedFunctionException.cs" />
     <Compile Include="Operations\BaseRepository\Exists.cs" />
     <Compile Include="Operations\BaseRepository\Sum.cs" />
     <Compile Include="Operations\BaseRepository\SumAll.cs" />
